LOWZERO/requirements.txt
git+https://github.com/teepark/python-lzf.git
pycryptodome